Moved CommanderProductionInfo helper class from Muriel to AIUtils.
